Update and rename ISSprOM 2019_4000.omap to ISSprOM 2019-2_4000.omap
Update following errata of ISSprOM2019-2
Modify somme colors to corresponding to" Map_Specifications_Printing_And_Colour_Definitions". This document must be updated.
I have modify the order of "White for railroad" to corresponding to the table => and i have modify the symbol "railroad" for it to work with this new order.
I assumed:
Dark green is (100 0 80 30) (describe in errata in ISOM2017-2)
Green 100% become (76 0 91 0) like in ISOM2017 because now dark green is back.